Relocate spos related transaction code
