+++ src/main/scala/scorex/transaction/TransactionParser.scala
@@ -7,6 +7,7 @@ import scorex.transaction.contract.{ChangeContractStatusTransaction, CreateContr
 import scorex.transaction.database.DbPutTransaction
 import scorex.transaction.lease.{LeaseCancelTransaction, LeaseTransaction}
 import vee.transaction.MintingTransaction
+import vee.transaction.spos.{ContendSlotsTransaction, ReleaseSlotsTransaction}
 
 import scala.util.{Failure, Try}
