+++ src/main/scala/vee/api/http/spos/ReleaseSlotsRequest.scala
@@ -1,15 +1,15 @@
-package scorex.api.http.spos
+package vee.api.http.spos
 
 import io.swagger.annotations.ApiModelProperty
 import play.api.libs.json.{Format, Json}
 
-case class ContendSlotsRequest(@ApiModelProperty(value = "Base58 encoded sender public key", required = true)
+case class ReleaseSlotsRequest(@ApiModelProperty(value = "Base58 encoded sender public key", required = true)
                               sender: String,
                               @ApiModelProperty(required = true)
                               slotids: Int,
                               @ApiModelProperty(required = true)
                               fee: Long)
 
-object ContendSlotsRequest {
-  implicit val aliasRequestFormat: Format[ContendSlotsRequest] = Json.format
+object ReleaseSlotsRequest {
+  implicit val aliasRequestFormat: Format[ReleaseSlotsRequest] = Json.format
 }
