+++ src/main/scala/vee/transaction/spos/ContendSlotsTransaction.scala
@@ -1,4 +1,4 @@
-package scorex.transaction
+package vee.transaction.spos
 
 import com.google.common.primitives.{Bytes, Longs, Ints}
 import com.wavesplatform.state2.ByteStr
@@ -7,18 +7,19 @@ import scorex.account._
 import scorex.crypto.EllipticCurveImpl
 import scorex.crypto.hash.FastCryptographicHash
 import scorex.transaction.TransactionParser._
+import scorex.transaction._
 
 import scala.util.{Failure, Success, Try}
 
 
-case class ReleaseSlotsTransaction private(sender: PublicKeyAccount,
+case class ContendSlotsTransaction private(sender: PublicKeyAccount,
                                            slotid: Int,
                                            fee: Long,
                                            timestamp: Long,
                                            signature: ByteStr)
   extends SignedTransaction {
 
-  override val transactionType: TransactionType.Value = TransactionType.ReleaseSlotsTransaction
+  override val transactionType: TransactionType.Value = TransactionType.ContendSlotsTransaction
 
   lazy val toSign: Array[Byte] = Bytes.concat(
     Array(transactionType.id.toByte),
@@ -30,6 +31,7 @@ case class ReleaseSlotsTransaction private(sender: PublicKeyAccount,
   override lazy val id: ByteStr = ByteStr(FastCryptographicHash(toSign))
 
   override lazy val json: JsObject = jsonBase() ++ Json.obj(
+    //need to write more data here?
     "slotid" -> slotid,
     "fee" -> fee,
     "timestamp" -> timestamp
@@ -40,9 +42,9 @@ case class ReleaseSlotsTransaction private(sender: PublicKeyAccount,
 
 }
 
-object ReleaseSlotsTransaction {
+object ContendSlotsTransaction {
 
-  def parseTail(bytes: Array[Byte]): Try[ReleaseSlotsTransaction] = Try {
+  def parseTail(bytes: Array[Byte]): Try[ContendSlotsTransaction] = Try {
     import EllipticCurveImpl._
     val sender = PublicKeyAccount(bytes.slice(0, KeyLength))
     val (slotBytes, slotidEnd) = (bytes.slice(KeyLength, KeyLength + SlotidLength), KeyLength + SlotidLength)
@@ -50,7 +52,7 @@ object ReleaseSlotsTransaction {
     val fee = Longs.fromByteArray(bytes.slice(slotidEnd, slotidEnd + 8))
     val timestamp = Longs.fromByteArray(bytes.slice(slotidEnd + 8, slotidEnd + 16))
     val signature = ByteStr(bytes.slice(slotidEnd + 16, slotidEnd + 16 + SignatureLength))
-    ReleaseSlotsTransaction
+    ContendSlotsTransaction
       .create(sender, slotid, fee, timestamp, signature)
       .fold(left => Failure(new Exception(left.toString)), right => Success(right))
   }.flatten
@@ -59,17 +61,17 @@ object ReleaseSlotsTransaction {
              slotid: Int,
              fee: Long,
              timestamp: Long,
-             signature: ByteStr): Either[ValidationError, ReleaseSlotsTransaction] =
+             signature: ByteStr): Either[ValidationError, ContendSlotsTransaction] =
     if (fee <= 0) {
       Left(ValidationError.InsufficientFee)
     } else {
-      Right(ReleaseSlotsTransaction(sender, slotid, fee, timestamp, signature))
+      Right(ContendSlotsTransaction(sender, slotid, fee, timestamp, signature))
     }
 
   def create(sender: PrivateKeyAccount,
              slotid: Int,
              fee: Long,
-             timestamp: Long): Either[ValidationError, ReleaseSlotsTransaction] = {
+             timestamp: Long): Either[ValidationError, ContendSlotsTransaction] = {
     create(sender, slotid, fee, timestamp, ByteStr.empty).right.map { unsigned =>
       unsigned.copy(signature = ByteStr(EllipticCurveImpl.sign(sender, unsigned.toSign)))
     }
