+++ src/main/scala/vee/transaction/spos/ReleaseSlotsTransaction.scala
@@ -1,4 +1,4 @@
-package scorex.transaction
+package vee.transaction.spos
 
 import com.google.common.primitives.{Bytes, Longs, Ints}
 import com.wavesplatform.state2.ByteStr
@@ -7,18 +7,19 @@ import scorex.account._
 import scorex.crypto.EllipticCurveImpl
 import scorex.crypto.hash.FastCryptographicHash
 import scorex.transaction.TransactionParser._
+import scorex.transaction._
 
 import scala.util.{Failure, Success, Try}
 
 
-case class ContendSlotsTransaction private(sender: PublicKeyAccount,
+case class ReleaseSlotsTransaction private(sender: PublicKeyAccount,
                                            slotid: Int,
                                            fee: Long,
                                            timestamp: Long,
                                            signature: ByteStr)
   extends SignedTransaction {
 
-  override val transactionType: TransactionType.Value = TransactionType.ContendSlotsTransaction
+  override val transactionType: TransactionType.Value = TransactionType.ReleaseSlotsTransaction
 
   lazy val toSign: Array[Byte] = Bytes.concat(
     Array(transactionType.id.toByte),
@@ -30,7 +31,6 @@ case class ContendSlotsTransaction private(sender: PublicKeyAccount,
   override lazy val id: ByteStr = ByteStr(FastCryptographicHash(toSign))
 
   override lazy val json: JsObject = jsonBase() ++ Json.obj(
-    //need to write more data here?
     "slotid" -> slotid,
     "fee" -> fee,
     "timestamp" -> timestamp
@@ -41,9 +41,9 @@ case class ContendSlotsTransaction private(sender: PublicKeyAccount,
 
 }
 
-object ContendSlotsTransaction {
+object ReleaseSlotsTransaction {
 
-  def parseTail(bytes: Array[Byte]): Try[ContendSlotsTransaction] = Try {
+  def parseTail(bytes: Array[Byte]): Try[ReleaseSlotsTransaction] = Try {
     import EllipticCurveImpl._
     val sender = PublicKeyAccount(bytes.slice(0, KeyLength))
     val (slotBytes, slotidEnd) = (bytes.slice(KeyLength, KeyLength + SlotidLength), KeyLength + SlotidLength)
@@ -51,7 +51,7 @@ object ContendSlotsTransaction {
     val fee = Longs.fromByteArray(bytes.slice(slotidEnd, slotidEnd + 8))
     val timestamp = Longs.fromByteArray(bytes.slice(slotidEnd + 8, slotidEnd + 16))
     val signature = ByteStr(bytes.slice(slotidEnd + 16, slotidEnd + 16 + SignatureLength))
-    ContendSlotsTransaction
+    ReleaseSlotsTransaction
       .create(sender, slotid, fee, timestamp, signature)
       .fold(left => Failure(new Exception(left.toString)), right => Success(right))
   }.flatten
@@ -60,17 +60,17 @@ object ContendSlotsTransaction {
              slotid: Int,
              fee: Long,
              timestamp: Long,
-             signature: ByteStr): Either[ValidationError, ContendSlotsTransaction] =
+             signature: ByteStr): Either[ValidationError, ReleaseSlotsTransaction] =
     if (fee <= 0) {
       Left(ValidationError.InsufficientFee)
     } else {
-      Right(ContendSlotsTransaction(sender, slotid, fee, timestamp, signature))
+      Right(ReleaseSlotsTransaction(sender, slotid, fee, timestamp, signature))
     }
 
   def create(sender: PrivateKeyAccount,
              slotid: Int,
              fee: Long,
-             timestamp: Long): Either[ValidationError, ContendSlotsTransaction] = {
+             timestamp: Long): Either[ValidationError, ReleaseSlotsTransaction] = {
     create(sender, slotid, fee, timestamp, ByteStr.empty).right.map { unsigned =>
       unsigned.copy(signature = ByteStr(EllipticCurveImpl.sign(sender, unsigned.toSign)))
     }
